Make script respect formatted timestamps
ffmpeg2gif would toss "invalid xx" errors if timestamps supplied
to either -s or -t were not plain integers.
This commit fixes that by allowing the passing of strings instead.
Timestamp format is, naturally, the same as for ffmpeg.
If you're used to passing integers, worry not - it still works.